chore: stricter input schema for pagination

+++ apps/website/src/server/trpc/router/admin/activity-feed.ts
@@ -7,7 +7,7 @@ export const adminActivityFeedRouter = router({
   getOutgoingWebhooks: superUserProcedure
     .input(
       z.object({
-        limit: z.number().min(1).max(100).nullish(),
+        limit: z.number().int().min(1).max(100).nullish(),
         cursor: z.string().cuid().nullish(),
       })
     )
